+++ playhouse/migrate.py
@@ -745,7 +745,8 @@ class SqliteMigrator(SchemaMigrator):
         new_column_defs = []
         new_column_names = []
         original_column_names = []
-        constraint_terms = ('foreign ', 'primary ', 'constraint ', 'check ')
+        constraint_terms = ('foreign ', 'primary ', 'constraint ', 'check ',
+                            'unique ')
 
         for column_def in column_defs:
             column_name, = self.column_name_re.match(column_def).groups()
@@ -790,7 +791,7 @@ class SqliteMigrator(SchemaMigrator):
 
         # Update the name of the new CREATE TABLE query.
         temp_table = table + '__tmp__'
-        rgx = re.compile(r'("?)%s("?)' % re.escape(table), re.I)
+        rgx = re.compile(r'("?)%s("?)\s*$' % re.escape(table), re.I)
         create = rgx.sub(
             r'\1%s\2' % temp_table,
             raw_create)